refactor: card front back refactoring
