+++ src/component/result/IDCard/common/character-info-box.tsx
