+++ src/component/result/IDCard/common/character-profile.tsx
@@ -2,6 +2,11 @@
 import styled from 'styled-components';
 
 import { FlexCenter } from '@/component/core/Flex';
+import {
+  ContainImage,
+  ProfileTagContainerStyled,
+  ProfileTagStyled,
+} from '@/component/result/IDCard/styles';
 import type { CharacterReturnType } from '@/utils/answer';
 
 export default function CharacterProfile({
@@ -12,11 +17,16 @@ export default function CharacterProfile({
   return (
     <CharacterWrapper>
       <CharacterMainImageWrapper>
-        <img src={character.image} width="90" height="68" alt="character" />
+        <ContainImage
+          src={character.image}
+          width="90"
+          height="68"
+          alt="character"
+        />
       </CharacterMainImageWrapper>
-      <BottomWrapper>
-        <Tag>{character.name}</Tag>
-      </BottomWrapper>
+      <ProfileTagContainerStyled>
+        <ProfileTagStyled>{character.name}</ProfileTagStyled>
+      </ProfileTagContainerStyled>
     </CharacterWrapper>
   );
 }
@@ -25,29 +35,6 @@ const CharacterWrapper = styled.div`
   display: flex;
   flex-direction: column;
   gap: 3px;
-`;
-
-const BottomWrapper = styled.div`
-  margin-top: 8px;
-  height: 20px;
-  position: relative;
-`;
-
-const Tag = styled.div`
-  position: absolute;
-  left: 50%;
-  transform: translateX(-50%);
-
-  width: fit-content;
-  font-family: 'GongGothicMedium';
-  font-weight: 700;
-  font-size: 10px;
-  padding: 0 5px;
-  line-height: 20px;
-
-  color: ${(props) => props.theme.colors.main2};
-  background: ${(props) => props.theme.colors.sub2};
-  border-radius: 10px;
 `;
 
 const CharacterMainImageWrapper = styled(FlexCenter)`
